refactor: make id always long

+++ src/main/java/io/goobi/vocabularyserver/repositories/FieldInstanceRepository.java
@@ -3,5 +3,5 @@ package io.goobi.vocabularyserver.repositories;
 import io.goobi.vocabularyserver.model.FieldInstance;
 import org.springframework.data.repository.CrudRepository;
 
-public interface FieldInstanceRepository extends CrudRepository<FieldInstance, Integer> {
+public interface FieldInstanceRepository extends CrudRepository<FieldInstance, Long> {
 }
