+++ src/main/java/io/goobi/vocabularyserver/repositories/FieldValidationRepository.java
@@ -3,5 +3,5 @@ package io.goobi.vocabularyserver.repositories;
 import io.goobi.vocabularyserver.model.FieldValidation;
 import org.springframework.data.repository.CrudRepository;
 
-public interface FieldValidationRepository extends CrudRepository<FieldValidation, Integer> {
+public interface FieldValidationRepository extends CrudRepository<FieldValidation, Long> {
 }
