+++ src/main/java/io/goobi/vocabularyserver/repositories/VocabularySchemaRepository.java
@@ -3,5 +3,5 @@ package io.goobi.vocabularyserver.repositories;
 import io.goobi.vocabularyserver.model.VocabularySchema;
 import org.springframework.data.repository.CrudRepository;
 
-public interface VocabularySchemaRepository extends CrudRepository<VocabularySchema, Integer> {
+public interface VocabularySchemaRepository extends CrudRepository<VocabularySchema, Long> {
 }
